clean up old simple pipeline
