+++ snuba/settings/settings_test.py
@@ -68,5 +68,3 @@ REDIS_CLUSTERS = {
     ]
 }
 VALIDATE_DATASET_YAMLS_ON_STARTUP = True
-USE_NEW_QUERY_PIPELINE_SAMPLE_RATE = 0.0
-TRY_NEW_QUERY_PIPELINE_SAMPLE_RATE = 1.0
