+++ snuba/web/query.py
@@ -1,13 +1,9 @@
 from __future__ import annotations
 
-import copy
 import logging
-import random
 from typing import Optional
 
 from snuba import environment
-from snuba import settings as snuba_settings
-from snuba import state
 from snuba.datasets.dataset import Dataset
 from snuba.datasets.factory import get_dataset_name
 from snuba.pipeline.query_pipeline import QueryPipelineResult
@@ -41,28 +37,17 @@ def _run_query_pipeline(
     concurrent_queries_gauge: Optional[Gauge] = None,
     force_dry_run: bool = False,
 ) -> QueryResult:
-    try_new_query_pipeline_rollout = state.get_float_config(
-        "try_new_query_pipeline", snuba_settings.TRY_NEW_QUERY_PIPELINE_SAMPLE_RATE
-    )
-    run_new_query_pipeline_rollout = state.get_float_config(
-        "run_new_query_pipeline", snuba_settings.USE_NEW_QUERY_PIPELINE_SAMPLE_RATE
-    )
-    try_new_query_pipeline = (
-        random.random() <= try_new_query_pipeline_rollout
-        if try_new_query_pipeline_rollout is not None
-        else False
-    )
-    run_new_query_pipeline = (
-        random.random() <= run_new_query_pipeline_rollout
-        if run_new_query_pipeline_rollout is not None
-        else False
-    )
     if isinstance(request.query, CompositeQuery):
         # New pipeline does not support composite queries yet.
-        try_new_query_pipeline = False
-        run_new_query_pipeline = False
-
-    if run_new_query_pipeline:
+        clickhouse_query = EntityAndStoragePipelineStage().execute(
+            QueryPipelineResult(
+                data=request,
+                query_settings=request.query_settings,
+                timer=timer,
+                error=None,
+            )
+        )
+    else:
         clickhouse_query = EntityProcessingStage().execute(
             QueryPipelineResult(
                 data=request,
@@ -72,44 +57,7 @@ def _run_query_pipeline(
             )
         )
         clickhouse_query = StorageProcessingStage().execute(clickhouse_query)
-    else:
-        if try_new_query_pipeline:
-            request_copy = copy.deepcopy(request)
-            compare_clickhouse_query = None
-            try:
-                compare_clickhouse_query = EntityProcessingStage().execute(
-                    QueryPipelineResult(
-                        data=request_copy,
-                        query_settings=request.query_settings,
-                        timer=timer,
-                        error=None,
-                    )
-                )
-                compare_clickhouse_query = StorageProcessingStage().execute(
-                    compare_clickhouse_query
-                )
-            except Exception as e:
-                logger.exception(e)
-        clickhouse_query = EntityAndStoragePipelineStage().execute(
-            QueryPipelineResult(
-                data=request,
-                query_settings=request.query_settings,
-                timer=timer,
-                error=None,
-            )
-        )
-        if try_new_query_pipeline:
-            if compare_clickhouse_query:
-                new_sql = str(compare_clickhouse_query.data)
-            else:
-                new_sql = ""
-            old_sql = str(clickhouse_query.data)
-            if new_sql != old_sql:
-                logger.warning(
-                    "New and old query pipeline sql doesn't match: Old: %s, New: %s",
-                    old_sql,
-                    new_sql,
-                )
+
     res = ExecutionStage(
         request.attribution_info,
         query_metadata=query_metadata,
